Port most of the std math lib to 32-bit IEEE 754
Since the basis for this library was use in a personal project, I opted
to remove functionality I didn't need rather than spending time porting
them. In a later commit to this repo they may be added back.

Also removed are the tests which have hard-coded constants.

The easiest work here is simply replacing all instances of 64-bit base
types (ie. flaot64, uint64, etc) with their 32-bit equivallences.
More tricky was make sure the bit shifting constants and masks were
correct as there are assumptions of size in many functions. In case, I
needed to hard code a bit mask.

+++ asin.go
@@ -16,9 +16,9 @@ package go-math32
 // Special cases are:
 //	Asin(±0) = ±0
 //	Asin(x) = NaN if x < -1 or x > 1
-func Asin(x float64) float64
+func Asin(x float32) float32
 
-func asin(x float64) float64 {
+func asin(x float32) float32 {
 	if x == 0 {
 		return x // special case
 	}
@@ -48,8 +48,8 @@ func asin(x float64) float64 {
 //
 // Special case is:
 //	Acos(x) = NaN if x < -1 or x > 1
-func Acos(x float64) float64
+func Acos(x float32) float32
 
-func acos(x float64) float64 {
+func acos(x float32) float32 {
 	return Pi/2 - Asin(x)
 }
